+++ src/__tests__/integration/cli.test.ts
@@ -1,10 +1,10 @@
 import path from 'path';
 import fs from 'fs';
 import child_process from 'child_process';
-import {promisify} from 'util';
-import {createTemporaryFs, isFile} from '../utils';
-import type {FsDefinition} from '../utils';
-import type {ChildProcess} from 'child_process';
+import { promisify } from 'util';
+import { createTemporaryFs, isFile } from '../utils';
+import type { FsDefinition } from '../utils';
+import type { ChildProcess } from 'child_process';
 
 const exec = promisify(child_process.exec);
 const readFile = promisify(fs.readFile);
@@ -22,8 +22,8 @@ async function _exec(
 ): Promise<ChildProcess> {
   const cliPath = path.join(PROJECT_ROOT, 'lib', 'cli.js');
   const cmd = `${cliPath} ${args.join(' ')}`;
-  const childProcess = await exec(cmd, {cwd});
-  const {stdout, stderr} = childProcess;
+  const childProcess = await exec(cmd, { cwd });
+  const { stdout, stderr } = childProcess;
   // eslint-disable-next-line no-console
   console.log(
     `EXEC ${cmd}\ncwd=${cwd}\nstdout:\n${stdout}\nstderr:\n${stderr}`
@@ -33,15 +33,16 @@ async function _exec(
 
 type TmpFsObject = {
   cwd: string;
-  exec: (path: Array<string>) => Promise<ChildProcess>;
+  exec: (a: Array<string>) => Promise<ChildProcess>;
 };
+
 type TmpFsCallback = (a: TmpFsObject) => unknown;
 
 async function createTmpFs(
   definition: FsDefinition,
   callback: TmpFsCallback
 ): Promise<void> {
-  const {cwd} = await createTemporaryFs({
+  const { cwd } = await createTemporaryFs({
     ...definition,
     './package.json': '{}',
   });
@@ -60,7 +61,7 @@ describe('cli', () => {
         './a.js': '',
         './modules.js': `import './a.js'`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['--recast.quote="double"', './a.js', './b.js']);
         expect(await readFileString(path.join(cwd, './modules.js'))).toEqual(
           `import "./b.js"`
@@ -76,7 +77,7 @@ describe('cli', () => {
         './modules.js': `import './a.js'`,
         './modules.d.js': `import './a.js'`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec([
           '--ignore-pattern="*.d.js"',
           '--ignore-pattern="bogus"',
@@ -118,7 +119,7 @@ require(\`llo \${y} a \u0000\`);
 require(x);
 require(x + '1');`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./a.js', './b.js']);
         expect(await readFileString(path.join(cwd, './modules.js'))).toEqual(`
 // dependency or built-in
@@ -156,7 +157,7 @@ import _3 from './a';
 import './a';
 proxyquire('./a', {});`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./a.js', './b.js']);
         expect(await readFileString(path.join(cwd, './modules.js'))).toEqual(`
 import('./b');
@@ -176,7 +177,7 @@ proxyquire('./b', {});`);
         './folder/': '',
         './a.js': `import './a1';`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./a.js', './folder/b.js']);
         expect(await readFileString(path.join(cwd, './folder/b.js'))).toEqual(
           `import '../a1';`
@@ -193,7 +194,7 @@ proxyquire('./b', {});`);
         './folder/': '',
         './a.js': `import './other/a1'; import './b';`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./a.js', './other/a1.js', './folder/']);
         expect(await readFileString(path.join(cwd, './folder/a.js'))).toEqual(
           `import './a1'; import '../b';`
@@ -208,7 +209,7 @@ proxyquire('./b', {});`);
         './source/level/a1.js': `import '../a2';`,
         './source/a2.js': '',
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./source', './target/']);
         expect(
           await readFileString(path.join(cwd, './target/level/a1.js'))
@@ -223,7 +224,7 @@ proxyquire('./b', {});`);
         './a.someWeirdExtension': '',
         './modules.js': `import './a.someWeirdExtension';`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./a.someWeirdExtension', './a.otherExt']);
         expect(await readFileString(path.join(cwd, './modules.js'))).toEqual(
           `import './a.otherExt';`
@@ -238,7 +239,7 @@ proxyquire('./b', {});`);
         './source-folder/a.sass': '@import "../b.sass";',
         './b.sass': '',
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./source-folder/*', '.']);
         expect(await readFileString(path.join(cwd, './a.sass'))).toEqual(
           '@import "./b.sass";'
@@ -254,7 +255,7 @@ proxyquire('./b', {});`);
         './source-folder/b.sass': '@import "../c.sass";',
         './c.sass': '',
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./source-folder/*', '.']);
         expect(await readFileString(path.join(cwd, './a.sass'))).toEqual(
           '@import "./b.sass";'
@@ -274,7 +275,7 @@ proxyquire('./b', {});`);
         './source-folder/b.sass': '@import "../c.sass";',
         './c.sass': '',
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./source-folder/*', '.']);
         expect(await readFileString(path.join(cwd, './a.pdf'))).toEqual(
           '%PDF- @import "./b.sass";'
@@ -292,7 +293,7 @@ proxyquire('./b', {});`);
         './a.someWeirdExtension': `@import ' ./modles.js ';`,
         './modles.js': '',
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./modles.js', './m.js']);
         expect(
           await readFileString(path.join(cwd, './a.someWeirdExtension'))
@@ -307,7 +308,7 @@ proxyquire('./b', {});`);
         './mod/index.js': '',
         './modules.js': `import './mod';`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./mod', './mod2']);
         expect(await readFileString(path.join(cwd, './modules.js'))).toEqual(
           `import './mod2';`
@@ -322,7 +323,7 @@ proxyquire('./b', {});`);
         './mod/index.js': '',
         './modules.js': `import './mod/index';`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./mod', './mod2']);
         expect(await readFileString(path.join(cwd, './modules.js'))).toEqual(
           `import './mod2/index';`
@@ -337,7 +338,7 @@ proxyquire('./b', {});`);
         './a.jsx': '',
         './modules.jsx': `import './a.jsx';`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./a.jsx', './b.jsx']);
         expect(await readFileString(path.join(cwd, './modules.jsx'))).toEqual(
           `import './b.jsx';`
@@ -353,7 +354,7 @@ proxyquire('./b', {});`);
         './a.mjs': '',
         './modules.mjs': `import './a.mjs';`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./a.mjs', './b.mjs']);
         expect(await readFileString(path.join(cwd, './modules.mjs'))).toEqual(
           `import './b.mjs';`
@@ -369,7 +370,7 @@ proxyquire('./b', {});`);
         './a.tsx': '',
         './modules.ts': `import './a.tsx';`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./a.tsx', './b.tsx']);
         expect(await readFileString(path.join(cwd, './modules.ts'))).toEqual(
           `import './b.tsx';`
@@ -387,7 +388,7 @@ proxyquire('./b', {});`);
         './c/': '',
         './modules.js': `import './a.js'; import './b';`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         await exec(['./a.js', './b.js', './c']);
         expect(await readFileString(path.join(cwd, './modules.js'))).toEqual(
           `import './c/a.js'; import './c/b';`
@@ -405,7 +406,7 @@ proxyquire('./b', {});`);
         './a/b/bar.js': '',
         './modules.js': `import './a/foo.js'; import './a/b/bar.js';`,
       },
-      async ({cwd, exec}) => {
+      async ({ cwd, exec }) => {
         const join = path.join.bind(path, cwd);
         await exec(['./a', './c/d']);
         expect(await readFileString(join('./modules.js'))).toEqual(
