+++ src/__tests__/options.test.ts
@@ -1,15 +1,15 @@
 import path from 'path';
-import {createMovePaths} from '../options';
-import {createTemporaryFs} from './utils';
+import { createMovePaths } from '../options';
+import { createTemporaryFs } from './utils';
 
 async function createFsAndMovePaths({
   sourcePaths,
   targetPath,
 }: {
-  sourcePaths: Array<string>;
+  sourcePaths: ReadonlyArray<string>;
   targetPath: string;
 }) {
-  const {cwd} = await createTemporaryFs({
+  const { cwd } = await createTemporaryFs({
     [targetPath]: '',
     ...sourcePaths.reduce((acc, s) => ((acc[s] = ''), acc), {}),
   });
@@ -18,12 +18,12 @@ async function createFsAndMovePaths({
     sourcePaths: sourcePaths.map((s) => join(s)),
     targetPath: join(targetPath),
   });
-  return {join, pathMap};
+  return { join, pathMap };
 }
 
 describe('normalize', () => {
   test('renames one file sourcePath to its targetPaths', async () => {
-    const {join, pathMap} = await createFsAndMovePaths({
+    const { join, pathMap } = await createFsAndMovePaths({
       sourcePaths: ['./foo.js'],
       targetPath: './bar.js',
     });
@@ -33,7 +33,7 @@ describe('normalize', () => {
   });
 
   test('renames one directory sourcePath to its targetPaths', async () => {
-    const {join, pathMap} = await createFsAndMovePaths({
+    const { join, pathMap } = await createFsAndMovePaths({
       sourcePaths: ['./a/s/d'],
       targetPath: './baz',
     });
@@ -43,7 +43,7 @@ describe('normalize', () => {
   });
 
   test('normalizes sourcePaths and targetPaths into a PathMap', async () => {
-    const {join, pathMap} = await createFsAndMovePaths({
+    const { join, pathMap } = await createFsAndMovePaths({
       sourcePaths: ['./foo.js', './a/bar.js', './a/s/d'],
       targetPath: './baz/',
     });
@@ -55,7 +55,7 @@ describe('normalize', () => {
   });
 
   test('normalizes directory targetPath into a PathMap', async () => {
-    const {join, pathMap} = await createFsAndMovePaths({
+    const { join, pathMap } = await createFsAndMovePaths({
       sourcePaths: ['./foo.js', './a/bar.js', './a/s/d'],
       targetPath: './c/',
     });
