--- src/index.js
@@ -1,4 +1,0 @@
-// @flow
-
-export { transform } from './runner';
-export { move } from './move';
