+++ src/log.ts
@@ -1,10 +1,10 @@
-import {basename, dirname, extname} from 'path';
-import {codeFrameColumns} from '@babel/code-frame';
+import { basename, dirname, extname } from 'path';
+import { codeFrameColumns } from '@babel/code-frame';
 import debug from 'debug';
 import * as chalk from 'chalk';
 import pkg from '../package.json';
-import type {Debugger} from 'debug';
-import type {File} from './transform';
+import type { Debugger } from 'debug';
+import type { File } from './transform';
 
 // Forces colors. Ideally I'd figure out a way to make colors work when
 // using execFile.
@@ -32,12 +32,18 @@ export function createDebug(filename: string): Debugger {
 type LogOptions = {
   file?: File;
   loc?: {
-    start: { line: number; column: number };
-    end: { line: number; column: number };
+    start: {
+      line: number;
+      column: number;
+    };
+    end: {
+      line: number;
+      column: number;
+    };
   };
 };
 
-function createFrame({file, loc}: LogOptions = {}): string {
+function createFrame({ file, loc }: LogOptions = {}): string {
   if (!file) {
     return '';
   }
