+++ src/path.ts
@@ -1,15 +1,15 @@
 import path from 'path';
 import fs from 'fs';
 import findUp from 'find-up';
-import {promisify} from 'util';
+import { promisify } from 'util';
 import ignore from 'ignore';
 import requireResolve from './requireResolve';
-import {createDebug, warn} from './log';
-import {JS_EXTENSIONS_DOTTED} from './options';
-import type {PathMap} from './options';
-import type {Context} from './transform';
+import { createDebug, warn } from './log';
+import { JS_EXTENSIONS_DOTTED } from './options';
+import type { PathMap } from './options';
+import type { Context } from './transform';
 
-const {hasOwnProperty} = Object.prototype;
+const { hasOwnProperty } = Object.prototype;
 const debug = createDebug(__filename);
 
 const readdir = promisify(fs.readdir);
@@ -71,7 +71,7 @@ function getAbsoluteImportSourcePath(
   context: Context,
   importSourcePath: string
 ): string {
-  const {file} = context;
+  const { file } = context;
   const absolutePath = path.resolve(path.dirname(file.path), importSourcePath);
   if (pathHasJSExtension(file.path)) {
     // It only makes sense to use requireResolve on JS files
@@ -148,7 +148,7 @@ export function updateSourcePath(
   context: Context,
   importSourcePath: string
 ): string {
-  const {file} = context;
+  const { file } = context;
 
   // absolute paths...
   // They are generaly not used "as-is", but there is a babel plugin that
@@ -166,7 +166,7 @@ export function updateSourcePath(
     return importSourcePath;
   }
 
-  const {options} = context;
+  const { options } = context;
   const absoluteImportSourcePath = getAbsoluteImportSourcePath(
     context,
     importSourcePath
@@ -216,8 +216,8 @@ function splitByFilter(
   array: Array<string>,
   filterFn: (a: string) => boolean
 ): [Array<string>, Array<string>] {
-  const truthyValues: Array<string> = [];
-  const falsyValues: Array<string> = [];
+  const truthyValues = [];
+  const falsyValues = [];
   array.forEach((value) => {
     const arr = filterFn(value) ? truthyValues : falsyValues;
     arr.push(value);
@@ -236,7 +236,7 @@ function pathHasJSExtension(_path: string): boolean {
 const IGNORED_FOLDERS = new Set(['node_modules', '.git', '.hg']);
 
 type FindAllOptions = {
-  ignorePattern: Array<string>;
+  ignorePattern: ReadonlyArray<string>;
 };
 
 // TODO: use `git ls-files` and `hg manifest` to consider ignored files.
@@ -258,11 +258,14 @@ function applyIgnoreFilter(paths, ignorePattern) {
 export async function findAllPathsCategorized(
   rootPath: string,
   options: FindAllOptions
-): Promise<{ js: Array<string>; others: Array<string> }> {
+): Promise<{
+  js: Array<string>;
+  others: Array<string>;
+}> {
   const allPaths = await findAllPaths(rootPath);
   const filteredPaths = applyIgnoreFilter(allPaths, options.ignorePattern);
   const [js, others] = splitByFilter(filteredPaths, pathHasJSExtension);
-  return {js, others};
+  return { js, others };
 }
 
 async function findAllPaths(
@@ -304,8 +307,8 @@ async function findAllPaths(
  */
 export async function expandDirectoryPaths(pathMap: PathMap): Promise<PathMap> {
   return await Object.keys(pathMap)
-    .map((sourcePath) => ({sourcePath, stat: stat(sourcePath)}))
-    .reduce(async (_acc, {sourcePath, stat}) => {
+    .map((sourcePath) => ({ sourcePath, stat: stat(sourcePath) }))
+    .reduce(async (_acc, { sourcePath, stat }) => {
       const acc = await _acc;
       if ((await stat).isDirectory()) {
         const allDescendants = await findAllPaths(sourcePath);
