+++ src/runner.ts
@@ -1,6 +1,6 @@
 import path from 'path';
-import {isBinaryFile} from 'isbinaryfile';
-import {createDebug, info} from './log';
+import { isBinaryFile } from 'isbinaryfile';
+import { createDebug, info } from './log';
 import {
   findAllPathsCategorized,
   findProjectPath,
@@ -9,9 +9,9 @@ import {
   writeFile,
   updateSourcePath,
 } from './path';
-import {validate, createMovePaths, DEFAULT, JS_EXTENSIONS} from './options';
-import type {MoveOptions, PathMap} from './options';
-import type {ParsedOptions} from './transform';
+import { validate, createMovePaths, DEFAULT, JS_EXTENSIONS } from './options';
+import type { MoveOptions, PathMap } from './options';
+import type { ParsedOptions } from './transform';
 import Runner from 'jscodeshift/src/Runner';
 
 const debug = createDebug(__filename);
@@ -21,7 +21,7 @@ type TransformOptions = {
   parser?: 'flow' | 'babylon' | 'babel';
   // see https://github.com/benjamn/recast/blob/master/lib/options.ts
   recastOptions?: any;
-  ignorePattern: Array<string>;
+  ignorePattern: ReadonlyArray<string>;
 };
 
 type NormalizedOptions = {
@@ -29,8 +29,10 @@ type NormalizedOptions = {
 } & TransformOptions;
 
 async function promiseObject(object: {
-  [key: string]: Promise<unknown> | unknown;
-}): Promise<{ [key: string]: unknown }> {
+  [x: string]: Promise<unknown> | unknown;
+}): Promise<{
+  [x: string]: unknown;
+}> {
   return await Promise.all(
     Object.entries(object).map((entry) => Promise.all(entry))
   ).then((entries) =>
@@ -38,7 +40,7 @@ async function promiseObject(object: {
       const [k, v] = entry;
       acc[k] = v;
       return acc;
-    }, {} as {[key: string]: unknown})
+    }, {})
   );
 }
 
@@ -46,8 +48,9 @@ async function genericTransform(
   paths: Array<string>,
   options: ParsedOptions
 ): Promise<void> {
-  for await (const {_path, content} of paths.map((p) =>
-    promiseObject({_path: p, content: readFile(p)})
+  // $FlowFixMe :shrug:
+  for await (const { _path, content } of paths.map((p) =>
+    promiseObject({ _path: p, content: readFile(p) })
   )) {
     const isFileBinary = await isBinaryFile(content);
     // Ignore binary files
@@ -69,7 +72,7 @@ async function genericTransform(
         // transform. It contains the values that are actually used by `updateSourcePath`.
         const context = {
           j: null,
-          file: {path: _path, source: ''},
+          file: { path: _path, source: '' },
           options,
         };
         return `${quote}${updateSourcePath(context, filePath.trim())}${quote}`;
@@ -83,13 +86,13 @@ async function genericTransform(
 export async function executeTransform(
   options: NormalizedOptions
 ): Promise<void> {
-  const {expandedPaths, ignorePattern} = options;
+  const { expandedPaths, ignorePattern } = options;
   debug('expandedPaths', JSON.stringify(expandedPaths, null, 2));
 
   const projectPath = await findProjectPath();
   info(`Detected project path: ${projectPath}`);
 
-  const recastOptions = {...DEFAULT.recast, ...options.recastOptions};
+  const recastOptions = { ...DEFAULT.recast, ...options.recastOptions };
   // I'm considering that when there are no recast options it's because the
   // user doesn't care much about these options, so we only log them when
   // any recast option is passed.
@@ -100,7 +103,7 @@ export async function executeTransform(
   const {
     js: allJSPaths,
     others: allOtherPaths,
-  } = await findAllPathsCategorized(projectPath, {ignorePattern});
+  } = await findAllPathsCategorized(projectPath, { ignorePattern });
   debug('Detected js paths', `\n  ${allJSPaths.join('\n  ')}`);
 
   const transformOptions: ParsedOptions = {
